Personal Expenses
Suppose that the total tax you pay on your job income is 14%.  After
paying the taxes, you spend 10% of your net income to buy new clothes
and other accessories for the next school year and 1% to buy school
supplies. After buying clothes and school supplies, you use 25% of the
remaining money to buy savings bonds. For each dollar you spend to buy
savings bonds, your parents spend $0.50 to buy additional savings bonds
for you. This program prompts the user to enter the pay rate for an hour
and the number of hours you worked each week.